TOMLReader is being a part of Siv3D just now
